+++ apps/docs/src/components/Inputs/checkbox-group/ErrorView.tsx
@@ -1,4 +1,5 @@
-import { CheckboxGroup, Grid, VBox } from "@elf-framework/ui";
+import React from "react";
+import { CheckboxGroup, Checkbox, Grid, VBox } from "@react-elf/ui";
 
 export function ErrorView() {
   return (
@@ -14,16 +15,14 @@ export function ErrorView() {
         <div style={{ display: "flex", flexDirection: "column", padding: 20 }}>
           <div>&nbsp;</div>
           <CheckboxGroup
-            direction="vertical"
-            value={[10]}
+            values={[10, 20]}
             variant="danger"
-            options={[
-              { value: 10, label: "Travel" },
-              { value: 20, label: "Music" },
-              { value: 30, label: "Shopping" },
-            ]}
-            onChange={(e, values) => console.log(values)}
-          />
+            onChange={(values) => console.log(values)}
+          >
+            <Checkbox value={10}>Travel</Checkbox>
+            <Checkbox value={20}>Music</Checkbox>
+            <Checkbox value={30}>Shopping</Checkbox>
+          </CheckboxGroup>
         </div>
       </Grid>
     </VBox>