+++ packages/@react-elf/checkbox/src/context.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { CheckboxGroupState } from "./hooks/useCheckboxGroupState";
+
+export const CheckboxGroupContext = createContext<CheckboxGroupState | null>(
+  null
+);