+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @react-elf/docs
 
+## 0.0.83
+
+### Patch Changes
+
+- modify checkbox group component
+
+- Updated dependencies []:
+  - @react-elf/ui@0.0.83
+
 ## 0.0.82
 
 ### Patch Changes